Generalise checkpointing for both model types
